changed ingredient amount from BigDecimal to Double

+++ src/main/java/guru/springframework/converters/IngredientCommandToIngredient.java
@@ -6,8 +6,6 @@ import org.springframework.core.convert.converter.Converter;
 import org.springframework.lang.Nullable;
 import org.springframework.stereotype.Component;
 
-import java.math.BigDecimal;
-
 /**
  * Created by jt on 6/21/17.
  */
@@ -29,7 +27,7 @@ public class IngredientCommandToIngredient implements Converter<IngredientComman
 
         final Ingredient ingredient = new Ingredient();
         ingredient.setId(source.getId());
-        ingredient.setAmount(BigDecimal.valueOf(source.getAmount()));
+        ingredient.setAmount(source.getAmount());
         ingredient.setDescription(source.getDescription());
         ingredient.setUom(uomConverter.convert(source.getUom()));
         return ingredient;
